POINTS OF INTEREST (POI) APPLICATION PROJECT:
Created a GPS tracking application by implementing a user permission seeking function which utilizes user's GPS Location through the internet.
Implemented the onLocationChanged override function using the latLonViewmodel which passes in the live location data.
Created a GpsViewModel that takes the latlon data which passes new input values to the latLonLiveData, then set an observer to the latLonViewModel liveData which updates the user's live location to the UI.
Integrated an osmdroid map withing the GPS compose app, added a marker as an overlay to the map within the osmdroid AndroidView function.
Implemented the AndoidView update for tracking of the current geoPoint. Added the android navigation compose library & initialized the NavHost to host the composables within the Compose hierachy.
Implemented the settings composable function to return a callback function that updates the MapComposable geoPoints.
Created the homeScreen composable layout to use boxwithconstraints, calculating the appropriate map layout within the homescreen by subtracting the settings button from the max screen height.
CURRENT TASK: the update button on the settings screen does nothing on click, needs further modification.