SQLITE DATABASE FEATURE:
Installed the necessary dependencies in the build.gradle file (including: room-tkx & ksp room-compiler).
Created four new classes:
The data entity: class representing our actual data.
The DAO or Data Access Object: A class providing access to the database from the rest of the code, containing a series of methods which perform given SQL statements.
A class representing the actual database as a whole.