NAVIGATION COMPONENTS FEATURE:
Added a top app bar to the app, containing a hamburger menu icon which, when clicked, should bring up a navigation drawer with Map and Settings entries.
Added a floating and bottom app navigation bar containing two icons: for Home and Add Poi.
Created an input composable for adding a new poi which takes the POI name, type and description from the user.
Current Issue: when any of the navigation icons is clicked nothing happens. The top app bar navigation needs to open display the available nav options when clicked & the bottom bar Add Poi icon needs to navigate to the add poi screen when clicked.