NAVIGATION FEATURE UPDATE 1.1:
PREVIOUS ISSUE FIXED: Implemented the NavigationDrawer within the Scaffold in the MainNavBar composable.
Removed the multiple instances of the navController, drawerState and coroutinScope variables created & created them once in the MainActivity's setContent which is then passed to the MainNavBar.
Modified each composable created to take in the innerPaddingValues as parameter which value is passed in from the Scaffold innerPadding as an argument when called within the NavHost.
Modified the AddPoi composable layout using the BoxWithConstraints feature.