ADD POI FEATURE UPDATE 2.0:
Created a poiList state variable & initialized the poiViemodel within the parent composable of the map composable which is then observed with the poiList variable passed into it.
Passed the poiList state variable to the map composable as an argument which is then used to create a marker for each poi in the list.
Created a addPoi function within the ViewModel class which takes in a new poi object and updates the poiListViewModel liveData.
Modified the AddPoi composable to use the addPoi function from the ViewModel to add the newPoi object input from the user also adding the homeMenu callback to the button onClick actions to redirect the user to the map in the homeScreen after successfully adding a new poi & its marker.