POI SAVING FEATURE:
Created a getter functon within the ViewModel that gets and returns the current list of POIs saved in the PoiViewModel.
Created a SavePoi composable in the main activity that gets the poiList from the viewModel and inserts them to the local SQLite database through the Button onClick action.
Also created a backStack / 'Go Back' button set to return to the homeScreen for now.
Created the SavePoi route composable within the NavHost for routing.