LOCAL POI DATABASE FEATURE UPDATE 1.1:
Modified the MapComposable by moving the poi and currList loops to the update function of the Android View.
Changed the map1 variable in the creation of the marker to the view variable which receives the Map returned by the AndroidView.
Created a Toast to display a success message when the SavePOI button is clicked and saved successfully.
Commented out the Toast in the onLocationChange override function.